Removed VS compiler warnings (double to float conversion, loss of precision).
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@5522 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/states_screens/grand_prix_lose.hpp
@@ -20,7 +20,7 @@ class GrandPrixLose : public GUIEngine::Screen, public GUIEngine::ScreenSingleto
     float m_sky_angle;
     
     /** Global evolution of time */
-    double m_global_time;
+    float m_global_time;
     
     irr::scene::IMeshSceneNode* m_garage;
